Insane commit. Finished the layout.
If you want ToolBar stuff, llook inside of SearchResultsPage.java
If you want searching algorithm stuff, look inside of SearchTask and SearchTestTool. SearchTestTool describes how the data is retrieved, and SearchTask actually
searches the data

GUI defines the main GUI for the app. This includes the search bars, the case-sensitive button, and the search button
The SearchResults class describes the panel that displays the search results